last and first n chars from string
